Sessions

+++ backend/internal/models/driver.go
@@ -18,35 +18,35 @@ type Driver struct {
 	// Current team
 	TeamID *uuid.UUID `gorm:"type:uuid" json:"team_id"`
 
-	DriverNumber int8 `gorm:"uniqueIndex" json:"driver_number"`
+	DriverNumber int16 `gorm:"uniqueIndex" json:"driver_number"`
 
 	// Current Season
-	Season                 int8  `gorm:"not null" json:"season"`
-	SeasonPosition         *int8 `gorm:"default:NULL" json:"season_position"`
-	SeasonPoints           int8  `gorm:"default:0" json:"season_points"`
-	SeasonGrandPrixRaces   int8  `gorm:"default:0" json:"season_grand_prix_races"`
-	SeasonGrandPrixPoints  int8  `gorm:"default:0" json:"season_grand_prix_points"`
-	SeasonGrandPrixWins    int8  `gorm:"default:0" json:"season_grand_prix_wins"`
-	SeasonGrandPrixPodiums int8  `gorm:"default:0" json:"season_grand_prix_podiums"`
-	SeasonGrandPrixPoles   int8  `gorm:"default:0" json:"season_grand_prix_poles"`
+	Season                 int16  `gorm:"not null" json:"season"`
+	SeasonPosition         *int16 `gorm:"default:NULL" json:"season_position"`
+	SeasonPoints           int16  `gorm:"default:0" json:"season_points"`
+	SeasonGrandPrixRaces   int16  `gorm:"default:0" json:"season_grand_prix_races"`
+	SeasonGrandPrixPoints  int16  `gorm:"default:0" json:"season_grand_prix_points"`
+	SeasonGrandPrixWins    int16  `gorm:"default:0" json:"season_grand_prix_wins"`
+	SeasonGrandPrixPodiums int16  `gorm:"default:0" json:"season_grand_prix_podiums"`
+	SeasonGrandPrixPoles   int16  `gorm:"default:0" json:"season_grand_prix_poles"`
 
-	SeasonSprintRaces   int8 `gorm:"default:0" json:"season_sprint_races"`
-	SeasonSprintPoints  int8 `gorm:"default:0" json:"season_sprint_points"`
-	SeasonSprintWins    int8 `gorm:"default:0" json:"season_sprint_wins"`
-	SeasonSprintPodiums int8 `gorm:"default:0" json:"season_sprint_podiums"`
-	SeasonSprintPoles   int8 `gorm:"default:0" json:"season_sprint_poles"`
+	SeasonSprintRaces   int16 `gorm:"default:0" json:"season_sprint_races"`
+	SeasonSprintPoints  int16 `gorm:"default:0" json:"season_sprint_points"`
+	SeasonSprintWins    int16 `gorm:"default:0" json:"season_sprint_wins"`
+	SeasonSprintPodiums int16 `gorm:"default:0" json:"season_sprint_podiums"`
+	SeasonSprintPoles   int16 `gorm:"default:0" json:"season_sprint_poles"`
 
 	// Career stats
-	GrandPrixEntered          int8    `gorm:"default:0" json:"grand_prix_entered"`
-	CareerPoints              int8    `gorm:"default:0" json:"career_points"`
-	CareerWins                int8    `gorm:"default:0" json:"career_wins"`
-	CareerPodiums             int8    `gorm:"default:0" json:"career_podiums"`
-	CareerPolePositions       int8    `gorm:"default:0" json:"career_pole_positions"`
-	CareerWorldChampionships  int8    `gorm:"default:0" json:"career_world_championships"`
-	CareerDNFs                int8    `gorm:"default:0" json:"career_dnfs"`
+	GrandPrixEntered          int16   `gorm:"default:0" json:"grand_prix_entered"`
+	CareerPoints              int16   `gorm:"default:0" json:"career_points"`
+	CareerWins                int16   `gorm:"default:0" json:"career_wins"`
+	CareerPodiums             int16   `gorm:"default:0" json:"career_podiums"`
+	CareerPolePositions       int16   `gorm:"default:0" json:"career_pole_positions"`
+	CareerWorldChampionships  int16   `gorm:"default:0" json:"career_world_championships"`
+	CareerDNFs                int16   `gorm:"default:0" json:"career_dnfs"`
 	CareerHighestRaceFinish   *string `gorm:"type:varchar(100)" json:"career_highest_race_finish"`
 	CareerHighestGridPosition *string `gorm:"type:varchar(100)" json:"career_highest_grid_position"`
-	CareerFastestLaps         int8    `gorm:"default:0" json:"career_fastest_laps"`
+	CareerFastestLaps         int16   `gorm:"default:0" json:"career_fastest_laps"`
 
 	// Metadata
 	CreatedAt time.Time `json:"created_at"`
